MSA-15249: review packaging

+++ docker-entrypoint.sh
@@ -21,18 +21,27 @@ fi
 cd / || exit 1
 
 if [[ -L /opt/fmc_repository/CommandDefinition/cisco-ios-ms ]]; then
+	SYMLINK_TARGET=$(readlink -f /opt/fmc_repository/CommandDefinition/cisco-ios-ms || true)
+	if [[ -n "$SYMLINK_TARGET" && -e "$SYMLINK_TARGET/.git" ]]; then
+		log_info "🌹 Skipping upgrade for fellow developer."
+		exit 0
+	fi
+	log_info "🦥 Removing symlink."
 	rm -f /opt/fmc_repository/CommandDefinition/cisco-ios-ms
-	log_info "🍄 Removing symlink."
 fi
 
 # for backward compatibility with old backend, move existing repository to CommandDefinition
 if [[ -e /opt/fmc_repository/cisco-ios-ms/.git ]]; then
 	log_info "🦖 Moving existing git repository to /opt/fmc_repository/CommandDefinition for backend compatibility."
 	mkdir -p /opt/fmc_repository/CommandDefinition
-	mv /opt/fmc_repository/cisco-ios-ms /opt/fmc_repository/CommandDefinition/
+	if [[ -e /opt/fmc_repository/CommandDefinition/cisco-ios-ms ]]; then
+		log_info "🍄 Removing existing destination directory before move."
+		rm -rf /opt/fmc_repository/CommandDefinition/cisco-ios-ms
+	fi
+	mv /opt/fmc_repository/cisco-ios-ms /opt/fmc_repository/CommandDefinition/cisco-ios-ms
 elif [[ -d /opt/fmc_repository/cisco-ios-ms ]]; then
 	log_info "🐞 Not a git repository. Removing the directory for backend compatibility."
-	rm -rf  /opt/fmc_repository/cisco-ios-ms/
+	rm -rf  /opt/fmc_repository/cisco-ios-ms
 fi
 
 if [[ -e /opt/fmc_repository/CommandDefinition/cisco-ios-ms/.git ]]; then
@@ -42,3 +51,5 @@ fi
 #tar --overwrite --no-same-owner -xf /home/ncuser/fmc-repository.tar.xz -I 'xz -T0' --checkpoint=1000 --checkpoint-action=ttyout='%{%Y-%m-%d %H:%M:%S}t⏳ \033[1;37m(%d sec)\033[0m: \033[1;32m#%u\033[0m, \033[0;33m%{}T\033[0m\r'
 tar --overwrite --no-same-owner -xf /home/ncuser/fmc-repository.tar.xz -I 'xz -T0' --checkpoint=1000 --checkpoint-action=echo='%{%Y-%m-%d %H:%M:%S}t⏳ \033[1;37m(%d sec)\033[0m: \033[1;32m#%u\033[0m, \033[0;33m%{}T\033[0m'
 echo "✅ Success ..."
+
+
